Crate wait_condition

Simple condition variables that are convenience wrappers around wait queues.

Structs

SatisfiedWaitCondition
A type wrapper that guarantees a given condition has been met before allowing one task to notify (wake up) other Tasks waiting on a WaitCondition. See the condition_satisfied() method.

WaitCondition
A condition variable that allows multiple Tasks to wait for a condition to be met, upon which other Tasks can notify them. This is effectively a convenience wrapper around WaitQueue::wait_until().

Trait aliases

WaitConditionFn
The closure type that can be used within a WaitCondition: a parameterless function that returns a bool indicating whether the condition is met.