Struct wait_condition::SatisfiedWaitCondition

pub struct SatisfiedWaitCondition<'wc, F: WaitConditionFn> { /* fields omitted */ }

A type wrapper that guarantees a given condition has been met before allowing one task to notify (wake up) other Tasks waiting on a WaitCondition. See the condition_satisfied() method.

Implementations

impl<'wc, F: WaitConditionFn> SatisfiedWaitCondition<'wc, F>

pub fn notify_one(&self) -> bool

Wake up a random Task that is waiting on this condition.

Return

returns Ok(true) if a Task was successfully woken up,

returns Ok(false) if there were no Tasks waiting.

pub fn notify_specific(&self, task_to_wakeup: &TaskRef) -> bool

Wake up a specific Task that is waiting on this condition.

Return

returns true if the given Task was waiting and was successfully woken up,

returns false if there was no such Task waiting.